add TableData

+++ MakeApple/Assets/Scripts/GameB/TableData.cs
@@ -7,25 +7,25 @@ public class TableData : MonoBehaviour
         return 10L + level * 5;
     }
 
-    public float GetBonusAttackPer(float startHeight, float normalFallingSpeed, float recordTime)
+    public double GetBonusAttackPer(float startHeight, float normalFallingSpeed, float recordTime)
     {
         if (normalFallingSpeed == 0)
         {
             Debug.LogError("[TableData] GetBonusAttackPer() normalFalling speed can't be zero.");
-            return 0f;
+            return 0d;
         }
 
         var bestTime = startHeight / (normalFallingSpeed * 2);
         var normalTime = startHeight / normalFallingSpeed;
 
         if (normalTime == bestTime)
-            return 0f;
+            return 0d;
 
         if (recordTime <= bestTime)
-            return 1f;
+            return 1d;
         else if (recordTime < normalTime)
-            return 1f - (recordTime - bestTime) / (normalTime - bestTime);
+            return 1d - (recordTime - bestTime) / (normalTime - bestTime);
 
-        return 0f;
+        return 0d;
     }
 }
